move more stuff to types/helpers files
