+++ src/components/addBlockModal/addBlockModalHelpers.ts
@@ -16,6 +16,8 @@
  */
 
 import { BLOCK_RESULT_COLUMN_COUNT } from "@/components/addBlockModal/addBlockModalConstants";
+import { RegistryId } from "@/core";
+import { ItemKeyInput } from "@/components/addBlockModal/addBlockModalTypes";
 
 export function getFlatArrayIndex({
   rowIndex,
@@ -26,4 +28,23 @@ export function getFlatArrayIndex({
 }): number {
   // Layout items in the grid left to right, top to bottom
   return rowIndex * BLOCK_RESULT_COLUMN_COUNT + columnIndex;
+}
+
+// We need to provide an item key because we reorder elements on search
+// (See: https://react-window.vercel.app/#/api/FixedSizeGrid)
+// Here, we use the brick id as the key, which is the "value" prop on the search result option
+export function getItemKey({
+  columnIndex,
+  data: { blockOptions },
+  rowIndex,
+}: ItemKeyInput): RegistryId | number {
+  const resultIndex = getFlatArrayIndex({ rowIndex, columnIndex });
+  // Number of bricks for the last Grid row could be less than the number of columns
+  // Returning the index here, ItemRenderer will render an empty cell
+  if (resultIndex >= blockOptions.length) {
+    return resultIndex;
+  }
+
+  // eslint-disable-next-line security/detect-object-injection -- index is a number
+  return blockOptions[resultIndex]?.value;
 }
